Removed remaining code smells (hopefully)
