+++ JAICore/jaicore-processes/src/main/java/jaicore/processes/CommandExecutionException.java
@@ -1,0 +1,10 @@
+package jaicore.processes;
+
+@SuppressWarnings("serial")
+public class CommandExecutionException extends Exception {
+
+	public CommandExecutionException(final Exception e) {
+		super(e);
+	}
+
+}
