+++ softwareconfiguration/hasco/src/main/java/hasco/model/ComponentUtil.java
@@ -12,6 +12,11 @@ import java.util.Random;
  */
 public class ComponentUtil {
 
+	private ComponentUtil () {
+
+		/* intentionally left blank. Just necessary to avoid explicit instantiation */
+	}
+
 	/**
 	 * This procedure returns a ComponentInstance of the given Component with default parameterization.
 	 * Note that required interfaces are not resolved.
