+++ softwareconfiguration/mlplan/src/example/java/de/upb/crc901/automl/mlplan/examples/MLPlanARFFExample.java
@@ -8,20 +8,9 @@ import java.util.concurrent.TimeUnit;
 
 import de.upb.crc901.mlplan.core.MLPlan;
 import de.upb.crc901.mlplan.core.MLPlanBuilder;
-import de.upb.crc901.mlplan.gui.outofsampleplots.OutOfSampleErrorPlotPlugin;
 import de.upb.crc901.mlplan.multiclass.wekamlplan.weka.model.MLPipeline;
-import hasco.gui.statsplugin.HASCOModelStatisticsPlugin;
 import jaicore.basic.TimeOut;
-import jaicore.graphvisualizer.plugin.graphview.GraphViewPlugin;
-import jaicore.graphvisualizer.plugin.nodeinfo.NodeInfoGUIPlugin;
-import jaicore.graphvisualizer.plugin.solutionperformanceplotter.SolutionPerformanceTimelinePlugin;
-import jaicore.graphvisualizer.window.AlgorithmVisualizationWindow;
 import jaicore.ml.WekaUtil;
-import jaicore.planning.hierarchical.algorithms.forwarddecomposition.graphgenerators.tfd.TFDNodeInfoGenerator;
-import jaicore.search.gui.plugins.rollouthistograms.SearchRolloutHistogramPlugin;
-import jaicore.search.model.travesaltree.JaicoreNodeInfoGenerator;
-import javafx.application.Platform;
-import javafx.embed.swing.JFXPanel;
 import weka.classifiers.Classifier;
 import weka.classifiers.evaluation.Evaluation;
 import weka.core.Instances;
@@ -43,13 +32,13 @@ public class MLPlanARFFExample {
 		MLPlan mlplan = new MLPlan(builder, split.get(0));
 		mlplan.setPortionOfDataForPhase2(0f);
 		mlplan.setLoggerName("mlplan");
-		mlplan.setTimeout(300, TimeUnit.SECONDS);
+		mlplan.setTimeout(60, TimeUnit.SECONDS);
 		mlplan.setNumCPUs(1);
 
-		new JFXPanel();
-		AlgorithmVisualizationWindow window = new AlgorithmVisualizationWindow(mlplan, new GraphViewPlugin(), new NodeInfoGUIPlugin<>(new JaicoreNodeInfoGenerator<>(new TFDNodeInfoGenerator())), new SearchRolloutHistogramPlugin<>(),
-				new SolutionPerformanceTimelinePlugin(), new HASCOModelStatisticsPlugin(), new OutOfSampleErrorPlotPlugin(split.get(0), split.get(1)));
-		Platform.runLater(window);
+		//		new JFXPanel();
+		//		AlgorithmVisualizationWindow window = new AlgorithmVisualizationWindow(mlplan, new GraphViewPlugin(), new NodeInfoGUIPlugin<>(new JaicoreNodeInfoGenerator<>(new TFDNodeInfoGenerator())), new SearchRolloutHistogramPlugin<>(),
+		//				new SolutionPerformanceTimelinePlugin(), new HASCOModelStatisticsPlugin(), new OutOfSampleErrorPlotPlugin(split.get(0), split.get(1)));
+		//		Platform.runLater(window);
 
 		try {
 			long start = System.currentTimeMillis();
